+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This also means you can write your own adapter and use any SIP library of your c
 Object has different properties!
 - **BREAKING**: `DEC112Specifics` constructor uses object `DEC112Config` instead of individual parameters for initialization.
 - **BREAKING**: Updating `Agent`'s heartbeat interval with invalid values will not result in an exception, but will rather log an error message.
+- **BREAKING**: `Agent`'s `debug` is now an object. This way, separate log-levels and handlers for in-library log messages and log messages created by the SIP adapter can be specified.
 - Improved build process, enabling better tree shaking for dependency `pidf-lo`.
 ### Added
 - Agent mode that can be used to indicate an app is running in the background.